requirements.txt
google-analytics-data==0.17.1
google-api-python-client==2.111.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.3